Move CaseDiffQueue to its own module
